style: refactored the project - made main.py cleaner
