+++ src/api/v1/__init__.py
@@ -3,6 +3,7 @@ from fastapi import APIRouter
 from src.api.v1.auth import router as auth_router
 from src.api.v1.calendar import router as calendar_router
 from src.api.v1.evaluations import router as evals_router
+from src.api.v1.frontend_routes import router as front_router
 from src.api.v1.meetings import router as meetings_router
 from src.api.v1.tasks import router as tasks_router
 from src.api.v1.teams import router as teams_router
@@ -14,3 +15,4 @@ main_router.include_router(teams_router)
 main_router.include_router(evals_router)
 main_router.include_router(meetings_router)
 main_router.include_router(calendar_router)
+main_router.include_router(front_router)
